+++ resources/assets/js/components/partials/discussionBoard.vue
@@ -1,11 +1,14 @@
 <template>
 <div :class="{'hidden': (activeTab != 'discussions')}" class="w-full">
   <create-discussion-form :resourceId="resource.id" :resourceType="resourceType" @close="closeCreateDiscussionForm" :form-shown="createDiscussionFormShown"></create-discussion-form>
+
+  <discussion-details v-if="discussionDetailsShown" :discussion="discussion" @close="closeDiscussionDetails"></discussion-details>
+
   <div class="text-center">
     <button @click="showCreateDiscussionForm" class="no-underline p-3 my-4 bg-white text-base text-teal rounded shadow">Create New Post</button>
   </div>
   <div class="flex flex-row flex-wrap justify-center items-start">
-    <div v-for="discussion in discussions" class="w-80 my-6 md:m-6 bg-white shadow-md flex flex-col items-center rounded cursor-pointer">
+    <div @click="showDiscussionDetails(index)" v-for="(discussion, index) in discussions" class="w-80 my-6 md:m-6 bg-white shadow-md flex flex-col items-center rounded cursor-pointer">
       <div class="bg-teal flex flex-col items-center w-full text-white rounded-t">
         <div class="w-10 h-10 flex-none py-4">
           <img :src="generateUrl(discussion.creator.avatar)" class="rounded-full w-10 h-10">
@@ -25,8 +28,12 @@
 
 <script>
 import createDiscussionForm from '../forms/createDiscussionForm.vue'
+import discussionDetails from './discussionDetails'
 export default {
-  components: {createDiscussionForm},
+  components: {
+    createDiscussionForm,
+    discussionDetails
+  },
   props: {
     resource: {
       required: true,
@@ -43,7 +50,9 @@ export default {
   },
   data: () => ({
     createDiscussionFormShown: false,
-    discussions: []
+    discussions: [],
+    discussion: {},
+    discussionDetailsShown: false
   }),
   methods: {
     showCreateDiscussionForm () {
@@ -68,6 +77,14 @@ export default {
           console.log(error.response.data.message)
         })
       }
+    },
+    showDiscussionDetails (index) {
+      this.discussion = this.discussions[index]
+      this.discussionDetailsShown = true
+    },
+    closeDiscussionDetails () {
+      this.discussionDetailsShown = false
+      this.discussion = {}
     }
   },
   watch: {