+++ tests/JS/setup.js
@@ -1,5 +1,5 @@
 Object.defineProperty(window, 'Laravel', {
-    value: {
-        csrfToken: Math.floor(Math.random() * 9999999999)
-    }
-});
+  value: {
+    csrfToken: Math.floor(Math.random() * 9999999999)
+  }
+})